requirements.txt
# ============================================================
# requirements.txt - Python Backend Dependencies
# Smart Agriculture Advisory Platform
# ============================================================

# ── Web Framework ─────────────────────────────────────────────
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1

# ── Database ──────────────────────────────────────────────────
PyMySQL==1.1.0
cryptography==41.0.7          # required by PyMySQL for SSL

# ── HTTP & API ────────────────────────────────────────────────
requests==2.31.0

# ── ML & Data Science ────────────────────────────────────────
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0            # for disease detection CNN
Pillow==10.1.0                # image processing

# ── Utilities ────────────────────────────────────────────────
python-dateutil==2.8.2
Werkzeug==3.0.1
gunicorn==21.2.0              # production WSGI server
python-dotenv==1.0.0          # load .env file
